tools/evals: split run.py into the ktw_evals package, one module per responsibility
run.py keeps the CLI and the driver notes in its docstring; the code moves
to ktw_evals/ - common, cases, workdir, drivers/ (one module per agent CLI,
registry in __init__), analysis, judge, results, runner, matrix. Every
documented command is unchanged. Definitions moved verbatim with their
comments; the retry loop is lifted out of main() into
runner.run_until_resolved(); the SKILL.md version regex that existed three
times is common.skill_version().

Checked before and after on the same host with the same two cases
(continuous-capture-basic, chestertons-fence-guard): identical record and
summary shape, both passing both times, matrix path smoke-tested with one
combination. The after-run caught two mistakes the import check did not
(TOOL_DIR one level too deep, a local shadowing the new helper).

Layout table in tools/evals/README.md, decision entry in context/evals.md
(new topic, indexed), CHANGELOG.
