+++ CHANGELOG.md
@@ -18,6 +18,8 @@ All notable changes to this project are documented here. Format follows [Keep a 
 
 - `references/repository-structure.md`: an **Evidence** paragraph alongside the Status and Type ones, stating what the field measures — how well a claim's *origin* is established, not whether it is true today. `confirmed` is good provenance (a maintainer said so, an authoritative source backs it), not a truth guarantee; whether it still holds is what Verification and Revisit when are for. The field definitions were only in `SKILL.md` rule 2, which names the three levels without saying that.
 
+- `tools/evals/run.py` (2,250 lines) is split into the `tools/evals/ktw_evals/` package: `common`, `cases`, `workdir`, `drivers/` (one module per agent CLI, registry in `__init__`), `analysis`, `judge`, `results`, `runner`, `matrix`. `run.py` stays the entry point — every documented command is unchanged, as is its module docstring with the driver notes. Mechanical: definitions moved verbatim with their comments, imports added, the retry loop lifted out of `main()` into `runner.run_until_resolved()`, and the SKILL.md version regex that existed three times is now `common.skill_version()`. Checked the way a mechanical change should be: the same two cases (`continuous-capture-basic`, `chestertons-fence-guard`) run before and after on the same host, identical record and summary shape, both passing both times — a check that earned its keep: it caught a path resolved one level too deep and a local variable shadowing the new helper, neither of which an import smoke test showed. Motivation is the next steps, not tidiness — deterministic per-case checks and separate activation/completion/judge counts each touch one module now instead of one file. Layout table in `tools/evals/README.md`.
+
 - The two shell snippets in `references/autostart.md` and `references/ci-linting.md` are fenced as ` ```sh ` instead of ` ```bash `. Same rendering; the reason is agent-skill-manager's install-time check, which labelled the package *High Risk* on the literal word `bash` alone. The remaining *Medium Risk* comes from the package containing links and is the floor for any skill that does. The security page now lists what every scanner that covers the skill reports (SkillsLLM, Socket, Snyk, Gen Agent Trust Hub, asm) and why the Snyk and Trust Hub medium findings — third-party content read during retrospective recovery and interviews — describe the skill's purpose and stay; `SECURITY.md` points there.
 
 - Project init wizard: the last question now asks how the skill should get loaded in future sessions, offering the three start paths (default: only when a developer asks); step 2 writes the entry-point section when that path is chosen; step 6's "leave the entry-point file alone" has that one exception.
