+++ tools/evals/ktw_evals/__init__.py
@@ -1,0 +1,1 @@
+"""Eval runner for the keep-the-why skill — the package behind tools/evals/run.py."""
