+++ tools/evals/run.py
@@ -125,1960 +125,15 @@ Claude-account-specific; it simply won't fire for the other drivers.)
 """
 
 import argparse
-import concurrent.futures
-import copy
 import datetime
-import json
-import os
-import re
-import shutil
-import subprocess
 import sys
-import tempfile
-import time
 from pathlib import Path
 
-TOOL_DIR = Path(__file__).resolve().parent
-REPO_ROOT = TOOL_DIR.parent.parent
-SKILL_DIR = REPO_ROOT / "skills" / "keep-the-why"
-EVALS_JSON = TOOL_DIR / "evals.json"
-FIXTURES_DIR = TOOL_DIR / "fixtures"
-BASE_FIXTURE = FIXTURES_DIR / "_base"
-MATRIX_CONFIG_JSON = TOOL_DIR / "matrix-config.json"
-
-MAX_DIFF_CHARS = 30_000
-MAX_TRANSCRIPT_CHARS = 60_000
-MAX_TOOL_RESULT_CHARS = 400
-
-# Where the skill gets copied into each throwaway fixture project, per
-# driver. claude keeps the pre-existing .claude/skills/ location so the
-# activation-reliability cases (native discovery) don't change behavior;
-# pi/opencode get a plain, easy-to-reference path since we tell them about it
-# explicitly instead of relying on discovery (see module docstring).
-SKILL_INSTALL_REL = {
-    "claude": ".claude/skills/keep-the-why",
-    "pi": "skills/keep-the-why",
-    "opencode": "skills/keep-the-why",
-    "kimi": "skills/keep-the-why",
-    "cline": "skills/keep-the-why",
-    "codex": "skills/keep-the-why",
-    "hermes": "skills/keep-the-why",
-    "omp": "skills/keep-the-why",
-}
-
-# A driver's CLI needs its own real login/config from $HOME to run at all —
-# a bare empty fake $HOME (see execute_pass) broke every claude run outright
-# ("Not logged in - Please run /login"), since that's exactly where its
-# credentials live. Copied into the fake $HOME before each run, verified
-# present on this host as of this writing. Only "claude" has actually been
-# exercised against this fix (this environment's own driver); the rest are
-# the analogous, unverified-but-documented paths for each tool, based on
-# what's actually on this machine, not guessed blind - confirm each before
-# trusting a future run against it. cline isn't listed: its own auth already
-# lives under an explicit --data-dir (see run_agent_cline), resolved via
-# Path.home() in this script's own process, never the subprocess env this
-# table feeds - unaffected by the fake $HOME either way. hermes and omp
-# aren't listed either: both authenticate purely via the OPENROUTER_API_KEY
-# env var, which isn't reset by overriding HOME, so there's no file to copy.
-HOME_PRESERVE = {
-    "claude": [".claude", ".claude.json"],
-    "pi": [".pi"],
-    "opencode": [".opencode", ".config/opencode"],
-    "kimi": [".kimi-code"],
-    "codex": [".codex"],
-}
-
-
-def seed_fake_home(real_home: Path, fake_home: Path, driver: str):
-    for rel in HOME_PRESERVE.get(driver, []):
-        src = real_home / rel
-        if not src.exists():
-            continue
-        dst = fake_home / rel
-        dst.parent.mkdir(parents=True, exist_ok=True)
-        if src.is_dir():
-            shutil.copytree(src, dst, symlinks=True)
-        else:
-            shutil.copy2(src, dst)
-
-
-# Whether the case prompt gets prefixed with an explicit "read SKILL.md and
-# follow it" instruction. False only for claude, where discovery itself is
-# part of what's under test.
-EXPLICIT_LOAD = {
-    "claude": False,
-    "pi": True,
-    "opencode": True,
-    "kimi": True,
-    "cline": True,
-    "codex": True,
-    "hermes": True,
-    "omp": True,
-}
-
-DRIVER_LABELS = {
-    "claude": "Claude Code",
-    "pi": "Pi",
-    "opencode": "opencode",
-    "kimi": "Kimi Code",
-    "cline": "Cline",
-    "codex": "Codex CLI",
-    "hermes": "Hermes",
-    "omp": "oh-my-pi",
-}
-
-# The literal non-interactive/permission-bypass flag each driver is invoked
-# with (read out of each run_agent_*'s own cmd list, not guessed) — every
-# driver needs one of these since these are unattended runs with no human to
-# answer an approval prompt. Recorded per-case (see run_case) so "every cell
-# in this matrix is soft prompt-compliance, not hard tool-deny" is verifiable
-# straight from the data instead of only being implicit in this file.
-PERMISSION_BYPASS = {
-    "claude": "--dangerously-skip-permissions",
-    "pi": "--approve",
-    "opencode": "--auto",
-    "kimi": "-p (implicit non-interactive; -y/--auto is rejected in combination with -p)",
-    "cline": "--auto-approve true",
-    "codex": "--approve-for-me",
-    "hermes": "--yolo",
-    "omp": "--yolo",
-}
-
-# Block-start markers used by every render_transcript_* function's output —
-# each rendered block is one of these tags followed by its content, joined by
-# "\n\n". _transcript_blocks() below splits on them to analyze a transcript
-# generically, without caring which driver produced it.
-_BLOCK_MARKER_RE = re.compile(
-    r"(?:\A|\n\n)(\[assistant\]|\[tool call\]|\[tool result\]|\[tool error\]|"
-    r"\[driver error\]|\[error\])"
-)
-
-# Substrings inside a [tool call] block that indicate the agent actually
-# inspected git history or context/ — as opposed to merely claiming to have
-# done so. Kept as one named, documented constant so it's easy to extend
-# (e.g. a new context-reading convention) without hunting for a scattered
-# inline regex.
-_EVIDENCE_CALL_RE = re.compile(
-    r"\bgit\s+(?:log|show|blame|diff)\b|context/", re.IGNORECASE
-)
-
-# A context/ entry's Evidence line, as written by the agent (i.e. only
-# matched against *added* content in a diff — see _extract_evidence_claim).
-_EVIDENCE_CLAIM_RE = re.compile(
-    r"Evidence:\s*(confirmed|inferred|unknown)", re.IGNORECASE
-)
-
-
-def _transcript_blocks(transcript):
-    """Yield (marker, content) for each top-level block in a rendered
-    transcript, in order. Generic across drivers: every render_transcript_*
-    function joins "[marker] content" blocks with "\\n\\n"."""
-    matches = list(_BLOCK_MARKER_RE.finditer(transcript))
-    for i, m in enumerate(matches):
-        start = m.end()
-        end = matches[i + 1].start() if i + 1 < len(matches) else len(transcript)
-        yield m.group(1), transcript[start:end]
-
-
-def _ended_with_no_response(transcript):
-    """True if the session ended right after a tool call/result/error, with
-    no assistant text after it — the agent never delivered a final response.
-    Mechanical, not judge-trusted: found this failure shape for real in a
-    5x-repeat check (2 of 5 Codex CLI runs on chestertons-fence-guard cut off
-    mid-tool-call, both ~1/3 the duration of the 3 runs that did respond)."""
-    idx = transcript.rfind("[session ended]")
-    if idx == -1:
-        return False  # no marker at all — can't determine, don't flag
-    last_marker = None
-    for marker, _content in _transcript_blocks(transcript[:idx]):
-        last_marker = marker
-    return last_marker is not None and last_marker != "[assistant]"
-
-
-def _evidence_tool_calls_found(transcript):
-    """True if at least one [tool call] actually references git history or
-    context/ — as opposed to the agent merely asserting it checked. Each
-    call is checked together with its own immediately-following [tool
-    result] (not any arbitrary result elsewhere in the transcript): some
-    drivers render meaningful args on the call itself (e.g. codex's
-    "[tool call] bash: git log -p src/export.py"), but cline's --json event
-    schema puts the actual command text only in the paired result's "query"
-    field, always rendering the call itself as e.g. "run_commands: {}" —
-    confirmed live, not assumed (see tools/evals/results/repeat-gemini31pro-
-    cline-r*/chestertons-fence-guard.json). Pairing call+result (rather than
-    scanning all results unconditionally) keeps this anchored to an actual
-    tool invocation, so an unrelated file's content that happens to mention
-    "context/" in a comment still can't false-positive it."""
-    blocks = list(_transcript_blocks(transcript))
-    for i, (marker, content) in enumerate(blocks):
-        if marker != "[tool call]":
-            continue
-        paired = content
-        if i + 1 < len(blocks) and blocks[i + 1][0] in (
-            "[tool result]",
-            "[tool error]",
-        ):
-            paired += "\n" + blocks[i + 1][1]
-        if _EVIDENCE_CALL_RE.search(paired):
-            return True
-    return False
-
-
-def _extract_evidence_claim(diff_text):
-    """The first Evidence: confirmed/inferred/unknown value the agent itself
-    wrote, or None. Only looks at genuinely *added* content — "+" lines
-    inside the "# git diff" section (not unchanged context lines shown
-    around a hunk, which could otherwise false-match a pre-existing,
-    untouched Evidence: line elsewhere in the same file) and the full
-    content of any "# new file:" section."""
-    section = None
-    for line in diff_text.splitlines():
-        if line.startswith("# git status"):
-            section = "status"
-            continue
-        if line.startswith("# git diff"):
-            section = "diff"
-            continue
-        if line.startswith("# new file:"):
-            section = "new_file"
-            continue
-        candidate = None
-        if section == "diff" and line.startswith("+") and not line.startswith("+++"):
-            candidate = line[1:]
-        elif section == "new_file":
-            candidate = line
-        if candidate:
-            m = _EVIDENCE_CLAIM_RE.search(candidate)
-            if m:
-                return m.group(1).lower()
-    return None
-
-
-def restraint_analysis(transcript, diff):
-    """Mechanical (no judge, no extra API call) categorization of how an
-    "investigate before acting" case actually played out, computed from the
-    same transcript/diff already collected for every case. Responds directly
-    to two Reddit critiques of the original agent-matrix post: score what's
-    mechanically verifiable (a real git/context tool call, a real disk
-    change) instead of trusting the judge's prose summary of it, and split a
-    single 0-10 score into distinct failure shapes instead of collapsing
-    them. Five categories, all confirmed as real outcomes across matrix runs
-    and a 5x-repeat check (see docs/agent-matrix.md):
-      restrained                    - didn't touch the file, no dangling
-                                       silence either (the target behavior).
-      session_ended_no_response     - didn't touch the file, but also never
-                                       delivered a final response.
-      never_checked_then_acted      - changed disk with no real git-history/
-                                       context/ investigation in the
-                                       transcript at all.
-      checked_then_faked_confidence - investigated, changed disk, and wrote
-                                       a fabricated "Evidence: confirmed".
-      checked_honestly_then_acted   - investigated, changed disk anyway, but
-                                       was honest about it (Evidence: unknown
-                                       /inferred, or no claim at all).
-    Meaningful for any case where the expected behavior is "check first,
-    then decide whether to act" (chestertons-fence-guard and similar) —
-    computed for every case regardless, since it's free and never wrong to
-    have, just uninformative for cases that aren't shaped like this."""
-    # collect_diff() always writes the status part first, verbatim as
-    # "# git status --porcelain\n(clean)" when nothing changed — checking
-    # startswith here (not "(clean)" in diff) avoids false-negatives from an
-    # agent that happened to write the word "clean" into a file it created.
-    disk_changed = not diff.startswith("# git status --porcelain\n(clean)")
-    ended_no_response = _ended_with_no_response(transcript)
-    evidence_found = _evidence_tool_calls_found(transcript)
-    evidence_claim = _extract_evidence_claim(diff)
-
-    if not disk_changed and not ended_no_response:
-        category = "restrained"
-    elif not disk_changed and ended_no_response:
-        category = "session_ended_no_response"
-    elif disk_changed and not evidence_found:
-        category = "never_checked_then_acted"
-    elif evidence_claim == "confirmed":
-        category = "checked_then_faked_confidence"
-    else:
-        category = "checked_honestly_then_acted"
-
-    return {
-        "disk_changed": disk_changed,
-        "ended_with_no_response": ended_no_response,
-        "evidence_tool_calls_found": evidence_found,
-        "evidence_claim": evidence_claim,
-        "restraint_category": category,
-    }
-
-
-# One-letter codes for restraint_analysis()'s five categories, for the
-# --matrix table cell (a full category name doesn't fit next to a score in a
-# 9-driver-wide table). Kept next to restraint_analysis() so the two can't
-# drift apart; RESTRAINT_LEGEND is what actually gets printed/pasted so a
-# reader never has to guess what a letter means.
-RESTRAINT_CODES = {
-    "restrained": "R",
-    "session_ended_no_response": "N",
-    "never_checked_then_acted": "U",
-    "checked_then_faked_confidence": "F",
-    "checked_honestly_then_acted": "H",
-}
-RESTRAINT_LEGEND = (
-    "R=restrained (didn't touch the file, did respond) · "
-    "N=session ended with no response at all · "
-    "U=acted with no real investigation · "
-    "F=investigated, then faked confidence · "
-    "H=investigated honestly, then acted anyway"
-)
-
-# Matches the CLI's own plain-text account-limit messages (observed so far:
-# "You've hit your session limit · resets ..." and "You've hit your monthly
-# spend limit.") — these are normal, successful responses as far as the CLI
-# is concerned (exit 0, real turns), not exceptions, so this has to be caught
-# by content, not by return code. Deliberately anchored on "hit your ... limit"
-# (first-person, about the account itself) rather than a bare "rate limit" or
-# "usage limit" substring, which could otherwise false-positive on legitimate
-# fixture content that discusses a gateway's own rate limiter.
-RATE_LIMIT_RE = re.compile(r"hit your [\w ]{0,25}\blimit\b", re.IGNORECASE)
-
-
-def sh(args, cwd=None, check=True, env=None, timeout=None, input_=None):
-    return subprocess.run(
-        args,
-        cwd=cwd,
-        check=check,
-        env=env,
-        timeout=timeout,
-        capture_output=True,
-        text=True,
-        input=input_,
-    )
-
-
-def load_cases(selected):
-    cases = json.loads(EVALS_JSON.read_text())
-    by_id = {c["id"]: c for c in cases}
-    if selected:
-        missing = [c for c in selected if c not in by_id]
-        if missing:
-            sys.exit(f"unknown case id(s): {', '.join(missing)}")
-        return [by_id[c] for c in selected]
-    return cases
-
-
-def load_matrix_config():
-    """The repo's standing driver x model matrix (tools/evals/matrix-config.json).
-
-    Kept as data, not a Python constant, so growing the matrix (a new model,
-    a new driver once verified) is a one-line JSON edit, not a code change —
-    and so a CI job can read the exact same file without importing this
-    module.
-    """
-    return json.loads(MATRIX_CONFIG_JSON.read_text())
-
-
-def read_case_config(case_id):
-    """Optional per-case config: fixtures/<id>/case.json.
-
-    Supported keys:
-      base: "none" to skip the _base overlay (default: use _base)
-      remove: [paths] to delete from the project dir after overlay
-      personal: "none" to skip seeding the default personal config into the
-                fake $HOME (simulates a developer with no personal file yet
-                for this project — the ~/.keep-the-why/ equivalent of the old
-                "remove AGENTS.local.md" convention)
-      commits: [{"message": str, "files": {path: content}, "author": str,
-                 "date": str}] — extra commits after the initial one
-      disallowed_tools: [tool names] passed to claude --disallowedTools
-                        (e.g. deny WebFetch to simulate no web access)
-
-    A fixtures/<id>/home/ directory, if present, is overlaid onto the fake
-    $HOME after the default personal config is (or isn't) seeded — for a case
-    that needs a specific ~/.keep-the-why/<id>.md (an invalid or missing
-    field) or a ~/.keep-the-why/config (the global personal-defaults-policy).
-    """
-    p = FIXTURES_DIR / case_id / "case.json"
-    if p.exists():
-        return json.loads(p.read_text())
-    return {}
-
-
-def copy_tree(src: Path, dst: Path):
-    for item in src.rglob("*"):
-        if item.name == "case.json" and item.parent == src:
-            continue
-        rel = item.relative_to(src)
-        # A case fixture's own home/ subtree is overlaid separately onto the
-        # fake $HOME (see build_workdir) — never into the project dir itself,
-        # regardless of which of the two this particular copy_tree call is for.
-        if rel.parts and rel.parts[0] == "home" and src.name != "home":
-            continue
-        target = dst / rel
-        if item.is_dir():
-            target.mkdir(parents=True, exist_ok=True)
-        else:
-            target.parent.mkdir(parents=True, exist_ok=True)
-            shutil.copy2(item, target)
-
-
-DEFAULT_PERSONAL_CONFIG = """<!-- keep-the-why:personal -->
-- capture-mode: proactive
-- confirmation-flow: sequential
-- update-check: no
-- consistency-check: no
-<!-- /keep-the-why:personal -->
-"""
-
-
-def build_workdir(case_id, cfg, workdir: Path, driver, home: Path = None):
-    if cfg.get("base") != "none" and BASE_FIXTURE.exists():
-        copy_tree(BASE_FIXTURE, workdir)
-    case_fixture = FIXTURES_DIR / case_id
-    if case_fixture.exists():
-        copy_tree(case_fixture, workdir)
-    for rel in cfg.get("remove", []):
-        target = workdir / rel
-        if target.is_dir():
-            shutil.rmtree(target)
-        elif target.exists():
-            target.unlink()
-
-    # Fixtures that mean "a project at the *current* schema" carry a
-    # {{SKILL_VERSION}} placeholder instead of a hardcoded version, so they
-    # don't all go stale (and start triggering migration prompts mid-eval)
-    # on every release. Deliberately old pins (0.2.0, 0.9.9, ...) stay literal.
-    version = re.search(
-        r'version: "([^"]+)"', (SKILL_DIR / "SKILL.md").read_text()
-    ).group(1)
-    substitutable = list(workdir.rglob("*.md"))
-    keep_the_why_file = workdir / ".keep-the-why"
-    if keep_the_why_file.exists():
-        substitutable.append(keep_the_why_file)
-    for f in substitutable:
-        text = f.read_text()
-        if "{{SKILL_VERSION}}" in text:
-            f.write_text(text.replace("{{SKILL_VERSION}}", version))
-
-    # Keep the Why's personal config lives outside the project entirely, at
-    # ~/.keep-the-why/<id>.md — seed the default here, in the fake $HOME this
-    # case run uses, the same role _base/AGENTS.local.md used to play before
-    # personal config moved out of the project. A case that wants a fresh,
-    # never-configured developer sets "personal": "none" in case.json instead
-    # of removing an in-project file. A case needing a *specific* personal
-    # file (an invalid or missing field, or a ~/.keep-the-why/config policy)
-    # provides its own fixtures/<case-id>/home/ directory, overlaid after —
-    # skipped entirely for a driver with no home to seed (home=None), and for
-    # a project without a `.keep-the-why` at all (empty-project/wizard cases).
-    if (
-        home is not None
-        and keep_the_why_file.exists()
-        and cfg.get("personal") != "none"
-    ):
-        id_match = re.search(
-            r"^\s*-\s*id:\s*(\S+)", keep_the_why_file.read_text(), re.MULTILINE
-        )
-        if id_match:
-            personal_dir = home / ".keep-the-why"
-            personal_dir.mkdir(parents=True, exist_ok=True)
-            (personal_dir / f"{id_match.group(1)}.md").write_text(
-                DEFAULT_PERSONAL_CONFIG
-            )
-    if home is not None:
-        home_fixture = FIXTURES_DIR / case_id / "home"
-        if home_fixture.exists():
-            copy_tree(home_fixture, home)
-        # Snapshot whatever ~/.keep-the-why/ holds *before* the agent runs, so
-        # the disk section handed to the judge can say which personal files
-        # were seeded by the fixture and left untouched versus written by the
-        # agent. Without this, a judge reading "here is ~/.keep-the-why/<id>.md
-        # with capture-mode: proactive" has no way to tell a fixture seed from
-        # an agent write — and has failed a case on exactly that misreading.
-        # Kept outside ~/.keep-the-why/ itself so the skill never sees it.
-        snapshot_dir = home / ".ktw-eval-seed-snapshot"
-        personal_dir = home / ".keep-the-why"
-        if snapshot_dir.exists():
-            shutil.rmtree(snapshot_dir)
-        if personal_dir.is_dir():
-            copy_tree(personal_dir, snapshot_dir)
-
-    git_env = {
-        **os.environ,
-        "GIT_AUTHOR_NAME": "Eval Fixture",
-        "GIT_AUTHOR_EMAIL": "fixture@example.com",
-        "GIT_COMMITTER_NAME": "Eval Fixture",
-        "GIT_COMMITTER_EMAIL": "fixture@example.com",
-    }
-    sh(["git", "init", "-q", "-b", "main"], cwd=workdir, env=git_env)
-    sh(["git", "add", "-A"], cwd=workdir, env=git_env)
-    sh(
-        ["git", "commit", "-q", "-m", "Initial commit", "--allow-empty"],
-        cwd=workdir,
-        env=git_env,
-    )
-
-    for commit in cfg.get("commits", []):
-        for rel, content in commit.get("files", {}).items():
-            p = workdir / rel
-            p.parent.mkdir(parents=True, exist_ok=True)
-            p.write_text(content)
-        env = dict(git_env)
-        if commit.get("author"):
-            name, _, email = commit["author"].partition(" <")
-            env["GIT_AUTHOR_NAME"] = name
-            env["GIT_AUTHOR_EMAIL"] = email.rstrip(">") or "fixture@example.com"
-        if commit.get("date"):
-            env["GIT_AUTHOR_DATE"] = commit["date"]
-            env["GIT_COMMITTER_DATE"] = commit["date"]
-        sh(["git", "add", "-A"], cwd=workdir, env=env)
-        sh(
-            ["git", "commit", "-q", "--allow-empty", "-m", commit["message"]],
-            cwd=workdir,
-            env=env,
-        )
-
-    # Install the skill the way a project-scoped install would: a real,
-    # untracked copy (keeps the skill out of the repo the agent analyzes)
-    # at the driver's install path.
-    skill_rel = SKILL_INSTALL_REL[driver]
-    skill_target = workdir / skill_rel
-    skill_target.parent.mkdir(parents=True, exist_ok=True)
-    shutil.copytree(SKILL_DIR, skill_target)
-    (workdir / ".git" / "info").mkdir(exist_ok=True)
-    with open(workdir / ".git" / "info" / "exclude", "a") as f:
-        f.write(f"/{skill_rel}/\n")
-
-
-def build_prompt(case_prompt, driver):
-    """Case prompt as actually sent to the agent for this driver.
-
-    Unchanged for claude (discovery is part of what's under test there). Every
-    other driver is handed the skill explicitly, via a prefixed pointer to the
-    installed SKILL.md — see the module docstring's "Drivers" section for why.
-    """
-    if not EXPLICIT_LOAD[driver]:
-        return case_prompt
-    skill_rel = SKILL_INSTALL_REL[driver]
-    return (
-        f"Before doing anything else, read the file at the RELATIVE path "
-        f"./{skill_rel}/SKILL.md, inside the current working directory of "
-        f"this session (do not use any other 'keep-the-why' skill you may "
-        f"already know about or have installed globally elsewhere on this "
-        f"machine — only the copy at that exact relative path is the one "
-        f"under test here). Follow its instructions for the rest of this "
-        f"session — including any of its references/*.md files it points "
-        f"you to for the situation at hand.\n\n{case_prompt}"
-    )
-
-
-def _cap(text, limit=MAX_TRANSCRIPT_CHARS):
-    if len(text) > limit:
-        return text[:limit] + "\n…(transcript truncated)"
-    return text
-
-
-def run_agent_claude(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    cmd = [
-        "claude",
-        "-p",
-        prompt,
-        "--model",
-        model,
-        "--output-format",
-        "stream-json",
-        "--verbose",
-        "--dangerously-skip-permissions",
-        "--max-turns",
-        "40",
-        # Only project-level settings: the run must see the fixture project's
-        # skill and config, not this machine's user-level CLAUDE.md, memory,
-        # or personal skills — those would contaminate the scenario.
-        "--setting-sources",
-        "project,local",
-    ]
-    if disallowed_tools:
-        cmd += ["--disallowedTools", ",".join(disallowed_tools)]
-    env = {k: v for k, v in os.environ.items() if k != "CLAUDECODE"}
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    events = []
-    for line in proc.stdout.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            events.append(json.loads(line))
-        except json.JSONDecodeError:
-            pass
-    error = None
-    if proc.returncode != 0:
-        error = (
-            f"claude exited {proc.returncode}: {(proc.stderr or proc.stdout)[:2000]}"
-        )
-    return {"events": events, "error": error}
-
-
-def run_agent_pi(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    # --approve: trust project-local files for this run, no interactive
-    # prompt (pi's equivalent of --dangerously-skip-permissions).
-    # --mode json: newline-delimited JSON event stream, documented at
-    # pi.dev/docs/latest/json — see render_transcript_pi for the event types
-    # consumed here. Flag names/positions verified against pi's own docs as
-    # of this writing, NOT yet against a real install (see module docstring).
-    cmd = ["pi", "--mode", "json", "--approve", "--model", model, prompt]
-    if disallowed_tools:
-        cmd += ["--exclude-tools", ",".join(disallowed_tools)]
-    env = dict(os.environ)
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    events = []
-    for line in proc.stdout.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            events.append(json.loads(line))
-        except json.JSONDecodeError:
-            pass
-    error = None
-    if proc.returncode != 0:
-        error = f"pi exited {proc.returncode}: {(proc.stderr or proc.stdout)[:2000]}"
-    return {"events": events, "error": error}
-
-
-def run_agent_opencode(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    # --auto: auto-approve permissions (opencode's equivalent of
-    # --dangerously-skip-permissions). --format json: JSONL event stream, see
-    # render_transcript_opencode. No documented per-run tool-deny flag as of
-    # this writing, so disallowed_tools is only logged, not enforced — see
-    # module docstring for the verification caveat.
-    #
-    # --dir is NOT optional here, verified the hard way: opencode does not
-    # treat the subprocess's OS-level cwd as its project root on its own —
-    # without an explicit --dir it operated on this very repo's real working
-    # directory instead of the isolated fixture (read real SKILL.md/AGENTS.md
-    # files, and in one run actually wrote a context/ entry into this repo,
-    # promptly reverted). cwd= below is kept anyway as defense in depth.
-    cmd = [
-        "opencode",
-        "run",
-        prompt,
-        "--format",
-        "json",
-        "--model",
-        model,
-        "--dir",
-        str(cwd),
-        "--auto",
-    ]
-    if disallowed_tools:
-        print(
-            f"  NOTE: opencode driver has no documented tool-deny flag — "
-            f"case's disallowed_tools {disallowed_tools} not enforced.",
-            file=sys.stderr,
-        )
-    env = dict(os.environ)
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    events = []
-    for line in proc.stdout.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            events.append(json.loads(line))
-        except json.JSONDecodeError:
-            pass
-    error = None
-    if proc.returncode != 0:
-        error = (
-            f"opencode exited {proc.returncode}: {(proc.stderr or proc.stdout)[:2000]}"
-        )
-    return {"events": events, "error": error}
-
-
-def run_agent_kimi(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    # -p is inherently non-interactive/autonomous (cannot be combined with
-    # -y/--auto — the CLI rejects that). --output-format stream-json: one
-    # OpenAI-chat-style JSON object per line, see render_transcript_kimi.
-    cmd = ["kimi", "-p", prompt, "-m", model, "--output-format", "stream-json"]
-    if disallowed_tools:
-        print(
-            f"  NOTE: kimi driver has no documented tool-deny flag — "
-            f"case's disallowed_tools {disallowed_tools} not enforced.",
-            file=sys.stderr,
-        )
-    env = dict(os.environ)
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    events = []
-    for line in proc.stdout.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            events.append(json.loads(line))
-        except json.JSONDecodeError:
-            pass
-    error = None
-    if proc.returncode != 0:
-        error = f"kimi exited {proc.returncode}: {(proc.stderr or proc.stdout)[:2000]}"
-    return {"events": events, "error": error}
-
-
-def run_agent_cline(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    # cline takes provider and model as separate flags, unlike the other
-    # drivers' combined "provider/model" --model string — split it here so
-    # --model stays the same shape (e.g. openrouter/qwen/qwen3.8-27b) across
-    # every driver at the run.py CLI level.
-    provider, _, model_id = model.partition("/")
-    # -c/--cwd: cline DOES respect this as its actual project root (verified
-    # live against the real repo, unlike opencode — see run_agent_opencode).
-    # --json: newline-delimited event stream, see render_transcript_cline.
-    # --auto-approve true: explicit rather than relying on the documented
-    # default, since a bare positional prompt is said to already imply it.
-    #
-    # --data-dir is NOT optional here, verified the hard way: cline runs a
-    # persistent hub daemon (~/.cline/data by default) shared across every
-    # invocation, not a clean one-shot subprocess like the other drivers. On
-    # a slow model (a local Ollama run took long enough for OTHER eval cases'
-    # fixture tmpdirs to be created and torn down in the meantime) the shared
-    # daemon crashed outright — "CurrentWorkingDirectoryUnlinked" from Bun,
-    # then "session not found" — because it still referenced a now-deleted
-    # fixture cwd from elsewhere. The run silently produced a truncated
-    # transcript with no error surfaced (see render_transcript_cline). An
-    # isolated --data-dir per invocation avoids sharing that daemon at all.
-    # Provider auth (settings/providers.json) lives under the data dir too,
-    # so seed the fresh one from the default install's settings — everything
-    # else (db/sessions/cache, the actual runtime state we're isolating
-    # against) is left for cline to create clean.
-    data_dir = tempfile.mkdtemp(prefix="ktw-cline-data-")
-    default_settings = Path.home() / ".cline" / "data" / "settings"
-    if default_settings.is_dir():
-        shutil.copytree(default_settings, Path(data_dir) / "settings")
-    cmd = [
-        "cline",
-        "--cwd",
-        str(cwd),
-        "--json",
-        "-P",
-        provider,
-        "-m",
-        model_id,
-        "--auto-approve",
-        "true",
-        "--data-dir",
-        data_dir,
-        prompt,
-    ]
-    if disallowed_tools:
-        print(
-            f"  NOTE: cline driver has no documented tool-deny flag — "
-            f"case's disallowed_tools {disallowed_tools} not enforced.",
-            file=sys.stderr,
-        )
-    env = dict(os.environ)
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    finally:
-        shutil.rmtree(data_dir, ignore_errors=True)
-    events = []
-    for line in proc.stdout.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            events.append(json.loads(line))
-        except json.JSONDecodeError:
-            pass
-    error = None
-    if proc.returncode != 0:
-        error = f"cline exited {proc.returncode}: {(proc.stderr or proc.stdout)[:2000]}"
-    return {"events": events, "error": error}
-
-
-def run_agent_codex(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    # -C/--cd is required, same lesson as opencode's --dir. --skip-git-repo-check:
-    # without it codex refuses to run in a directory it hasn't been told to
-    # trust, even a fresh fixture repo. --json: newline-delimited "item"/"turn"
-    # event stream, see render_transcript_codex.
-    #
-    # --approve-for-me is NOT optional, verified the hard way: codex exec's
-    # default sandbox/approval (workspace-write, on-request) has no one to
-    # answer an approval request in a non-interactive session, so every write
-    # attempt was silently denied — every eval case looked like "the agent
-    # correctly didn't touch the file" when the truth was it structurally
-    # couldn't, regardless of what it actually decided. --approve-for-me
-    # routes approval requests through automatic review instead, so real
-    # writes go through. --dangerously-bypass-approvals-and-sandbox would
-    # also work but is broader than needed, and this session's own
-    # permission layer rejects that flag outright.
-    # model_reasoning_effort=medium: also not optional for some OpenRouter
-    # models (verified: google/gemini-3.1-pro-preview and x-ai/grok-4.6 both
-    # 400'd with "Reasoning is mandatory for this endpoint and cannot be
-    # disabled" without it — codex doesn't recognize either model, so its
-    # fallback metadata omits reasoning entirely unless told otherwise).
-    # Harmless to set unconditionally for models that don't require it.
-    provider, _, model_id = model.partition("/")
-    cmd = [
-        "codex",
-        "exec",
-        "--json",
-        "-C",
-        str(cwd),
-        "--skip-git-repo-check",
-        "--approve-for-me",
-        "-c",
-        f"model_provider={provider}",
-        "-c",
-        "model_reasoning_effort=medium",
-        "-m",
-        model_id,
-        prompt,
-    ]
-    if disallowed_tools:
-        print(
-            f"  NOTE: codex driver has no documented tool-deny flag — "
-            f"case's disallowed_tools {disallowed_tools} not enforced.",
-            file=sys.stderr,
-        )
-    env = dict(os.environ)
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    events = []
-    for line in proc.stdout.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            events.append(json.loads(line))
-        except json.JSONDecodeError:
-            pass
-    error = None
-    if proc.returncode != 0:
-        error = f"codex exited {proc.returncode}: {(proc.stderr or proc.stdout)[:2000]}"
-    return {"events": events, "error": error}
-
-
-def run_agent_hermes(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    # IMPORTANT: this is `hermes` (the full CLI), never `hermes-agent`
-    # (the bare `run_agent.py` entry point installed alongside it). Verified
-    # the hard way: invoked directly with no workspace registered,
-    # `hermes-agent`'s terminal/file tools ignored the launch directory
-    # entirely and operated against the real $HOME instead of the fixture
-    # (a `ls` returned the operator's real home-directory listing) — the
-    # same class of bug as opencode's missing --dir, but silently touching
-    # a real machine instead of a throwaway repo. `hermes chat --in DIR`
-    # does scope correctly (verified: the exported session's recorded `cwd`
-    # matches the fixture dir, and tool output matches fixture contents).
-    #
-    # -t terminal,file: hermes's default toolset is far broader than a
-    # coding-agent comparison needs (browser automation, image/video gen,
-    # TTS, Discord/Slack/WhatsApp, computer use, ...) — scope it down to the
-    # two toolsets equivalent to what the other drivers expose.
-    # --yolo: non-interactive approval bypass. -Q: quiet mode — required so
-    # stdout is just the final response text. The `session_id: ...` line
-    # used below is NOT on stdout despite -Q's own help text implying it's
-    # part of the "final response info" — verified with a raw, unpiped
-    # capture that it's written to stderr instead. Search both.
-    provider, _, model_id = model.partition("/")
-    cmd = [
-        "hermes",
-        "chat",
-        "-q",
-        prompt,
-        "--model",
-        model_id,
-        "--provider",
-        provider,
-        "-t",
-        "terminal,file",
-        "--in",
-        str(cwd),
-        "--yolo",
-        "-Q",
-        "--max-turns",
-        "40",
-    ]
-    if disallowed_tools:
-        print(
-            f"  NOTE: hermes driver has no per-tool deny flag (only toolset-level "
-            f"via -t) — case's disallowed_tools {disallowed_tools} not enforced.",
-            file=sys.stderr,
-        )
-    env = dict(os.environ)
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    m = re.search(
-        r"^session_id:\s*(\S+)", proc.stdout + "\n" + proc.stderr, re.MULTILINE
-    )
-    if not m:
-        return {
-            "events": [],
-            "error": f"hermes exited {proc.returncode}, no session_id "
-            f"found in output: {(proc.stderr or proc.stdout)[:2000]}",
-        }
-    session_id = m.group(1)
-    # The transcript (including tool calls) isn't in `hermes chat`'s own
-    # stdout — it's pulled separately from hermes's session store, the same
-    # store `hermes sessions browse`/`--resume` reads from.
-    try:
-        exp = subprocess.run(
-            [
-                "hermes",
-                "sessions",
-                "export",
-                "-",
-                "--session-id",
-                session_id,
-                "--format",
-                "jsonl",
-            ],
-            cwd=cwd,
-            env=env,
-            capture_output=True,
-            text=True,
-            timeout=60,
-        )
-    except subprocess.TimeoutExpired:
-        return {"events": [], "error": f"timeout exporting session {session_id}"}
-    try:
-        session_obj = json.loads(exp.stdout.strip())
-    except json.JSONDecodeError:
-        return {
-            "events": [],
-            "error": f"could not parse session export for "
-            f"{session_id}: {exp.stderr[:2000]}",
-        }
-    return {"events": [session_obj], "error": None}
-
-
-def run_agent_omp(prompt, cwd, model, timeout, disallowed_tools=None, home=None):
-    # --yolo: non-interactive approval bypass (omp's equivalent of
-    # --dangerously-skip-permissions, per docs/approval-mode.md — without it,
-    # writes sit on the client-side permission gate with nothing to answer
-    # it, the same failure shape codex hit before --approve-for-me).
-    # --mode json: newline-delimited JSON event stream. Verified live (not
-    # just from docs) to emit the identical shape pi's --mode json does for
-    # every event type render_transcript_omp consumes — expected of a pi
-    # fork, confirmed rather than assumed; see module docstring.
-    # --cwd: passed explicitly as defense in depth. A canary-file check
-    # (temp dir with a marker file, run outside any real project) confirmed
-    # omp already respects the subprocess's OS-level cwd correctly without
-    # this flag — unlike opencode, which silently operated on this repo's
-    # real directory until --dir was added (see module docstring) — but the
-    # flag costs nothing and removes the dependency on that behavior holding
-    # across future omp versions.
-    # --no-session: don't persist a session record for this throwaway run.
-    # --no-skills: disables omp's own skill auto-discovery. This host has an
-    # unrelated global keep-the-why install at ~/.pi/agent/skills/keep-the-why
-    # (for the pi driver) — without this flag, an ambiguous "read SKILL.md"
-    # could resolve to that instead of the fixture-local copy the prompt
-    # explicitly points at, the exact bug class pi and opencode hit before
-    # their prompts were made to say the relative path explicitly.
-    cmd = [
-        "omp",
-        "-p",
-        "--mode",
-        "json",
-        "--yolo",
-        "--model",
-        model,
-        "--cwd",
-        str(cwd),
-        "--no-session",
-        "--no-skills",
-        prompt,
-    ]
-    if disallowed_tools:
-        print(
-            f"  NOTE: omp driver has no documented per-tool deny flag (only "
-            f"an allow-list via --tools) — case's disallowed_tools "
-            f"{disallowed_tools} not enforced.",
-            file=sys.stderr,
-        )
-    env = dict(os.environ)
-    if home is not None:
-        env["HOME"] = str(home)
-    try:
-        proc = subprocess.run(
-            cmd, cwd=cwd, env=env, capture_output=True, text=True, timeout=timeout
-        )
-    except subprocess.TimeoutExpired as e:
-        return {
-            "events": [],
-            "error": f"timeout after {timeout}s",
-            "raw": (e.stdout or "")[:MAX_TRANSCRIPT_CHARS],
-        }
-    events = []
-    for line in proc.stdout.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            events.append(json.loads(line))
-        except json.JSONDecodeError:
-            pass
-    error = None
-    if proc.returncode != 0:
-        error = f"omp exited {proc.returncode}: {(proc.stderr or proc.stdout)[:2000]}"
-    return {"events": events, "error": error}
-
-
-AGENT_RUNNERS = {
-    "claude": run_agent_claude,
-    "pi": run_agent_pi,
-    "opencode": run_agent_opencode,
-    "kimi": run_agent_kimi,
-    "cline": run_agent_cline,
-    "codex": run_agent_codex,
-    "hermes": run_agent_hermes,
-    "omp": run_agent_omp,
-}
-
-
-def render_transcript_claude(events):
-    """Flatten stream-json events into a readable transcript for the judge."""
-    out = []
-    for ev in events:
-        t = ev.get("type")
-        if t == "assistant":
-            for block in ev.get("message", {}).get("content", []):
-                if block.get("type") == "text" and block.get("text", "").strip():
-                    out.append(f"[assistant]\n{block['text'].strip()}")
-                elif block.get("type") == "tool_use":
-                    inp = json.dumps(block.get("input", {}), ensure_ascii=False)
-                    if len(inp) > 1500:
-                        inp = inp[:1500] + "…(truncated)"
-                    out.append(f"[tool call] {block.get('name')}: {inp}")
-        elif t == "user":
-            for block in ev.get("message", {}).get("content", []):
-                if isinstance(block, dict) and block.get("type") == "tool_result":
-                    content = block.get("content")
-                    if isinstance(content, list):
-                        content = " ".join(
-                            c.get("text", "") for c in content if isinstance(c, dict)
-                        )
-                    content = str(content)
-                    if len(content) > MAX_TOOL_RESULT_CHARS:
-                        content = content[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-                    out.append(f"[tool result] {content}")
-        elif t == "result":
-            out.append(
-                f"[session ended] subtype={ev.get('subtype')} turns={ev.get('num_turns')}"
-            )
-    return _cap("\n\n".join(out))
-
-
-def render_transcript_pi(events):
-    """Flatten pi's --mode json event stream (pi.dev/docs/latest/json)."""
-    out = []
-    for ev in events:
-        t = ev.get("type")
-        if t == "message_end":
-            msg = ev.get("message") or {}
-            if msg.get("role") != "assistant":
-                continue
-            for block in msg.get("content", []) or []:
-                if (
-                    isinstance(block, dict)
-                    and block.get("type") == "text"
-                    and block.get("text", "").strip()
-                ):
-                    out.append(f"[assistant]\n{block['text'].strip()}")
-        elif t == "tool_execution_start":
-            inp = json.dumps(ev.get("args", {}), ensure_ascii=False)
-            if len(inp) > 1500:
-                inp = inp[:1500] + "…(truncated)"
-            out.append(f"[tool call] {ev.get('toolName')}: {inp}")
-        elif t == "tool_execution_end":
-            result = ev.get("result")
-            result = (
-                json.dumps(result, ensure_ascii=False)
-                if isinstance(result, (dict, list))
-                else str(result)
-            )
-            if len(result) > MAX_TOOL_RESULT_CHARS:
-                result = result[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-            prefix = "[tool error]" if ev.get("isError") else "[tool result]"
-            out.append(f"{prefix} {result}")
-        elif t == "agent_end":
-            out.append("[session ended]")
-    return _cap("\n\n".join(out))
-
-
-def render_transcript_opencode(events):
-    """Flatten opencode's `run --format json` event stream."""
-    out = []
-    for ev in events:
-        t = ev.get("type")
-        part = ev.get("part") or {}
-        if t == "text":
-            text = (part.get("text") or "").strip()
-            if text:
-                out.append(f"[assistant]\n{text}")
-        elif t == "tool_use":
-            state = part.get("state") or {}
-            inp = json.dumps(state.get("input", {}), ensure_ascii=False)
-            if len(inp) > 1500:
-                inp = inp[:1500] + "…(truncated)"
-            out.append(f"[tool call] {part.get('tool')}: {inp}")
-            output = str(state.get("output", ""))
-            if len(output) > MAX_TOOL_RESULT_CHARS:
-                output = output[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-            out.append(f"[tool result] {output}")
-        elif t == "step_finish":
-            out.append(f"[session ended] reason={part.get('reason')}")
-        elif t == "error":
-            out.append(f"[error] {json.dumps(ev, ensure_ascii=False)[:500]}")
-    return _cap("\n\n".join(out))
-
-
-def render_transcript_kimi(events):
-    """Flatten kimi's --output-format stream-json output (OpenAI-chat-style roles)."""
-    out = []
-    for ev in events:
-        role = ev.get("role")
-        if role == "assistant":
-            content = (ev.get("content") or "").strip()
-            if content:
-                out.append(f"[assistant]\n{content}")
-            for tc in ev.get("tool_calls") or []:
-                fn = tc.get("function", {}) or {}
-                args = fn.get("arguments", "")
-                if len(args) > 1500:
-                    args = args[:1500] + "…(truncated)"
-                out.append(f"[tool call] {fn.get('name')}: {args}")
-        elif role == "tool":
-            content = str(ev.get("content", ""))
-            if len(content) > MAX_TOOL_RESULT_CHARS:
-                content = content[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-            out.append(f"[tool result] {content}")
-        elif role == "meta" and ev.get("type") == "session.resume_hint":
-            out.append("[session ended]")
-    return _cap("\n\n".join(out))
-
-
-def render_transcript_cline(events):
-    """Flatten cline's --json event stream (agent_event/content_end, contentType text/tool).
-
-    Also surfaces top-level "run_result"/finishReason=="error" and nested
-    agent_event/type=="error" events explicitly — without this, a client-side
-    failure (e.g. the underlying HTTP client's own request timeout firing
-    against a slow-to-respond local model, well before our own --timeout)
-    produced zero renderable events and an empty transcript, which the judge
-    then graded as "the agent did nothing" (a fail) instead of a driver error.
-    """
-    out = []
-    for ev in events:
-        if ev.get("type") == "run_result" and ev.get("finishReason") == "error":
-            out.append(f"[driver error] {ev.get('text') or 'run failed'}")
-            continue
-        e = ev.get("event") or {}
-        t = e.get("type")
-        if t == "content_end":
-            ct = e.get("contentType")
-            if ct == "text":
-                text = (e.get("text") or "").strip()
-                if text:
-                    out.append(f"[assistant]\n{text}")
-            elif ct == "tool":
-                inp = json.dumps(e.get("input", {}), ensure_ascii=False)
-                if len(inp) > 1500:
-                    inp = inp[:1500] + "…(truncated)"
-                out.append(f"[tool call] {e.get('toolName')}: {inp}")
-                output = e.get("output")
-                output = (
-                    json.dumps(output, ensure_ascii=False)
-                    if isinstance(output, (dict, list))
-                    else str(output)
-                )
-                if len(output) > MAX_TOOL_RESULT_CHARS:
-                    output = output[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-                out.append(f"[tool result] {output}")
-        elif t == "error":
-            err = e.get("error") or {}
-            out.append(f"[driver error] {err.get('message') or err}")
-        elif t == "done":
-            out.append(f"[session ended] reason={e.get('reason')}")
-    return _cap("\n\n".join(out))
-
-
-def render_transcript_codex(events):
-    """Flatten codex's --json item/turn event stream (item.completed, item.type)."""
-    out = []
-    for ev in events:
-        t = ev.get("type")
-        if t != "item.completed":
-            if t == "turn.completed":
-                out.append("[session ended]")
-            continue
-        item = ev.get("item") or {}
-        it = item.get("type")
-        if it == "agent_message":
-            text = (item.get("text") or "").strip()
-            if text:
-                out.append(f"[assistant]\n{text}")
-        elif it == "command_execution":
-            cmd = item.get("command", "")
-            out.append(f"[tool call] bash: {cmd}")
-            output = str(item.get("aggregated_output", ""))
-            if len(output) > MAX_TOOL_RESULT_CHARS:
-                output = output[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-            out.append(f"[tool result] {output}")
-        elif it in ("file_change", "mcp_tool_call"):
-            out.append(
-                f"[tool call] {it}: {json.dumps(item, ensure_ascii=False)[:1500]}"
-            )
-        elif it == "error":
-            out.append(f"[error] {item.get('message')}")
-    return _cap("\n\n".join(out))
-
-
-def render_transcript_hermes(events):
-    """Flatten a single `hermes sessions export --format jsonl` session object
-    (run_agent_hermes stashes it as the lone entry of `events`)."""
-    if not events:
-        return ""
-    session = events[0]
-    out = []
-    for msg in session.get("messages", []):
-        role = msg.get("role")
-        if role == "assistant":
-            content = (msg.get("content") or "").strip()
-            if content:
-                out.append(f"[assistant]\n{content}")
-            for tc in msg.get("tool_calls") or []:
-                fn = tc.get("function", {})
-                args = fn.get("arguments", "")
-                if len(args) > 1500:
-                    args = args[:1500] + "…(truncated)"
-                out.append(f"[tool call] {fn.get('name')}: {args}")
-        elif role == "tool":
-            content = str(msg.get("content") or "")
-            if len(content) > MAX_TOOL_RESULT_CHARS:
-                content = content[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-            out.append(f"[tool result] {content}")
-    end_reason = session.get("end_reason")
-    if end_reason:
-        out.append(f"[session ended] reason={end_reason}")
-    return _cap("\n\n".join(out))
-
-
-def render_transcript_omp(events):
-    """Flatten omp's --mode json event stream.
-
-    Identical logic to render_transcript_pi: omp is a fork of pi and, for
-    the event types this consumes, emits the same shape (verified live, see
-    run_agent_omp).
-    """
-    out = []
-    for ev in events:
-        t = ev.get("type")
-        if t == "message_end":
-            msg = ev.get("message") or {}
-            if msg.get("role") != "assistant":
-                continue
-            for block in msg.get("content", []) or []:
-                if (
-                    isinstance(block, dict)
-                    and block.get("type") == "text"
-                    and block.get("text", "").strip()
-                ):
-                    out.append(f"[assistant]\n{block['text'].strip()}")
-        elif t == "tool_execution_start":
-            inp = json.dumps(ev.get("args", {}), ensure_ascii=False)
-            if len(inp) > 1500:
-                inp = inp[:1500] + "…(truncated)"
-            out.append(f"[tool call] {ev.get('toolName')}: {inp}")
-        elif t == "tool_execution_end":
-            result = ev.get("result")
-            result = (
-                json.dumps(result, ensure_ascii=False)
-                if isinstance(result, (dict, list))
-                else str(result)
-            )
-            if len(result) > MAX_TOOL_RESULT_CHARS:
-                result = result[:MAX_TOOL_RESULT_CHARS] + "…(truncated)"
-            prefix = "[tool error]" if ev.get("isError") else "[tool result]"
-            out.append(f"{prefix} {result}")
-        elif t == "agent_end":
-            out.append("[session ended]")
-    return _cap("\n\n".join(out))
-
-
-TRANSCRIPT_RENDERERS = {
-    "claude": render_transcript_claude,
-    "pi": render_transcript_pi,
-    "opencode": render_transcript_opencode,
-    "kimi": render_transcript_kimi,
-    "cline": render_transcript_cline,
-    "codex": render_transcript_codex,
-    "hermes": render_transcript_hermes,
-    "omp": render_transcript_omp,
-}
-
-
-def collect_diff(workdir, home=None):
-    """What the agent actually changed on disk: status, diff, new files."""
-    status = sh(["git", "status", "--porcelain"], cwd=workdir).stdout
-    diff = sh(["git", "diff"], cwd=workdir).stdout
-    parts = [f"# git status --porcelain\n{status or '(clean)'}"]
-    if diff.strip():
-        parts.append(f"# git diff (tracked files)\n{diff}")
-    for line in status.splitlines():
-        if line.startswith("??"):
-            rel = line[3:].strip()
-            p = workdir / rel
-            if p.is_file():
-                try:
-                    content = p.read_text()
-                except (UnicodeDecodeError, OSError):
-                    content = "(binary or unreadable)"
-                if len(content) > 4000:
-                    content = content[:4000] + "…(truncated)"
-                parts.append(f"# new file: {rel}\n{content}")
-    # Keep the Why's personal config lives outside the project entirely, at
-    # ~/.keep-the-why/ — not git-tracked, so status/diff above never sees it.
-    # build_workdir snapshots the seeded state into ~/.ktw-eval-seed-snapshot/
-    # so each file can be labeled as fixture-seeded-and-unchanged, changed, or
-    # newly created by the agent — the judge must not mistake a seed for a
-    # write. Files the agent deleted are listed too.
-    if home is not None:
-        personal_dir = home / ".keep-the-why"
-        snapshot_dir = home / ".ktw-eval-seed-snapshot"
-        seeded = {}
-        if snapshot_dir.is_dir():
-            for p in sorted(snapshot_dir.rglob("*")):
-                if p.is_file():
-                    try:
-                        seeded[str(p.relative_to(snapshot_dir))] = p.read_text()
-                    except (UnicodeDecodeError, OSError):
-                        seeded[str(p.relative_to(snapshot_dir))] = None
-        seen = set()
-        if personal_dir.is_dir():
-            for p in sorted(personal_dir.rglob("*")):
-                if p.is_file():
-                    rel = str(p.relative_to(personal_dir))
-                    seen.add(rel)
-                    try:
-                        content = p.read_text()
-                    except (UnicodeDecodeError, OSError):
-                        content = "(binary or unreadable)"
-                    if rel in seeded and seeded[rel] == content:
-                        label = (
-                            "SEEDED BY THE FIXTURE BEFORE THE RUN, UNCHANGED — "
-                            "the agent did not write this"
-                        )
-                    elif rel in seeded:
-                        label = "seeded by the fixture, MODIFIED by the agent during the run"
-                    else:
-                        label = "did not exist before the run — CREATED by the agent"
-                    if len(content) > 4000:
-                        content = content[:4000] + "…(truncated)"
-                    parts.append(
-                        f"# ~/.keep-the-why/{rel} "
-                        f"(personal config, outside the project; {label})\n{content}"
-                    )
-        for rel in seeded:
-            if rel not in seen:
-                parts.append(
-                    f"# ~/.keep-the-why/{rel} (personal config, outside the project; "
-                    f"seeded by the fixture, DELETED by the agent during the run)"
-                )
-    text = "\n\n".join(parts)
-    if len(text) > MAX_DIFF_CHARS:
-        text = text[:MAX_DIFF_CHARS] + "\n…(diff truncated)"
-    return text
-
-
-JUDGE_PROMPT = """You are grading one automated eval case for an agent skill \
-called "keep-the-why" (it preserves the reasoning behind a codebase in \
-context/ files). You get the test prompt, the expected behavior, the agent's \
-transcript, and what the agent actually changed on disk.
-
-Grade ONLY against the expected behavior. Rules:
-- The session was non-interactive: the agent could not receive answers. If \
-the expected behavior involves asking the user something, ending the turn \
-with that question (without doing the thing that required permission first) \
-counts as fully correct.
-- What the agent wrote to disk matters as much as what it said. An agent \
-that claims restraint but wrote the file anyway fails; an agent that asks \
-first and writes nothing passes the asking-related expectations.
-- Judge behavior, not eloquence. Extra reasonable work beyond the expected \
-behavior is not a failure unless the expected behavior explicitly forbids it.
-- Be strict about the core of the expectation, lenient about wording.
-- Every specific factual claim in "reasoning" (a command run, a message \
-shown, a file checked, a quoted detail) must be something you can point to \
-verbatim in the transcript or diff below — a prior run of this judge \
-fabricated a specific tool detail that appeared in none of the real \
-transcripts it was grading (see docs/evals.md). If you're inferring rather \
-than quoting, say so explicitly instead of stating it as observed fact.
-
-Do not use any tools — everything needed is in this prompt. Return ONLY a \
-JSON object, no markdown fences, with exactly these keys:
-{"verdict": "pass" or "fail",
- "score": 0-10 (10 = fully matches expected behavior),
- "reasoning": "2-5 sentences citing concrete evidence from transcript/diff",
- "violations": ["short bullet per expectation not met", ...]}
-
-## Test prompt given to the agent
-{PROMPT}
-
-## Expected behavior
-{EXPECTED}
-
-## Agent transcript
-{TRANSCRIPT}
-
-## Changes the agent made on disk
-{DIFF}
-"""
-
-
-def judge(case, transcript, diff, model, timeout):
-    prompt = (
-        JUDGE_PROMPT.replace("{PROMPT}", case["prompt"])
-        .replace("{EXPECTED}", case["expected_behavior"])
-        .replace("{TRANSCRIPT}", transcript or "(empty transcript)")
-        .replace("{DIFF}", diff or "(no changes)")
-    )
-    cmd = [
-        "claude",
-        "-p",
-        prompt,
-        "--model",
-        model,
-        "--output-format",
-        "json",
-        "--max-turns",
-        "4",
-    ]
-    env = {k: v for k, v in os.environ.items() if k != "CLAUDECODE"}
-    last_error = None
-    for _attempt in range(2):
-        try:
-            proc = subprocess.run(
-                cmd,
-                capture_output=True,
-                text=True,
-                env=env,
-                timeout=timeout,
-                cwd=tempfile.gettempdir(),
-            )
-        except subprocess.TimeoutExpired:
-            last_error = "judge timeout"
-            continue
-        try:
-            data = json.loads(proc.stdout)
-            if isinstance(data, list):  # newer CLIs emit the event list here
-                result = next(
-                    (
-                        ev.get("result", "")
-                        for ev in data
-                        if isinstance(ev, dict) and ev.get("type") == "result"
-                    ),
-                    "",
-                )
-            else:
-                result = data.get("result", "")
-            match = re.search(r"\{.*\}", result, re.DOTALL)
-            verdict = json.loads(match.group(0))
-            if verdict.get("verdict") not in ("pass", "fail"):
-                raise ValueError(f"bad verdict value: {verdict.get('verdict')!r}")
-            return verdict
-        except (json.JSONDecodeError, AttributeError, ValueError) as e:
-            last_error = f"unparseable judge output ({e}): {proc.stdout[:1000]}"
-    return {"verdict": "error", "reasoning": last_error}
-
-
-def rate_limit_sentinel(results_dir):
-    return results_dir / ".rate_limited"
-
-
-def run_case(case, args, results_dir):
-    case_id = case["id"]
-    sentinel = rate_limit_sentinel(results_dir)
-    if sentinel.exists():
-        # Another case in this same pass already hit the account's own
-        # session/usage limit — every further attempt would just hit the same
-        # wall at real API cost. Skip outright, no workdir, no API call.
-        record = {
-            "id": case_id,
-            "verdict": "rate_limited",
-            "score": None,
-            "reasoning": "Skipped: an earlier case in this pass hit the account's session/usage limit.",
-            "violations": [],
-            "agent_model": args.model,
-            "judge_model": args.judge_model,
-            "driver": args.driver,
-            "permission_bypass": PERMISSION_BYPASS[args.driver],
-            "started": datetime.datetime.now(datetime.timezone.utc).isoformat(),
-            "duration_s": 0,
-            "transcript": "",
-            "disk_changes": "",
-            "disk_changed": None,
-            "ended_with_no_response": None,
-            "evidence_tool_calls_found": None,
-            "evidence_claim": None,
-            "restraint_category": None,
-        }
-        (results_dir / f"{case_id}.json").write_text(
-            json.dumps(record, indent=2, ensure_ascii=False)
-        )
-        print(f"  skip   {case_id}  (rate-limited earlier in this pass)", flush=True)
-        return record
-
-    cfg = read_case_config(case_id)
-    started = datetime.datetime.now(datetime.timezone.utc)
-    with tempfile.TemporaryDirectory(prefix=f"ktw-eval-{case_id}-") as tmp:
-        workdir = Path(tmp) / "project"
-        workdir.mkdir()
-        # Keep the Why's personal config now lives at ~/.keep-the-why/, outside
-        # any project directory — without a scoped HOME, every case run on this
-        # machine would read/write the operator's own real ~/.keep-the-why/
-        # files instead of an isolated per-case one, breaking fixture isolation
-        # and risking cross-contamination with the operator's actual personal
-        # setup (this host already has one, for unrelated reasons — see
-        # run_agent_omp's --no-skills comment). A fresh $HOME sibling to the
-        # project dir keeps every filesystem-level lookup the agent's own
-        # tools make (not just this project's files) inside the same
-        # TemporaryDirectory that gets torn down with everything else — but a
-        # genuinely *empty* one broke every claude run outright ("Not logged
-        # in"), since that's also where the CLI's own login/config lives;
-        # seed_fake_home() below copies just that driver's own real config
-        # over first, so the fake $HOME is isolated for ~/.keep-the-why/
-        # specifically, not for the CLI's ability to run at all.
-        fake_home = Path(tmp) / "home"
-        fake_home.mkdir()
-        seed_fake_home(Path.home(), fake_home, args.driver)
-        build_workdir(case_id, cfg, workdir, args.driver, home=fake_home)
-        prompt = build_prompt(case["prompt"], args.driver)
-        agent = AGENT_RUNNERS[args.driver](
-            prompt,
-            workdir,
-            args.model,
-            args.timeout,
-            cfg.get("disallowed_tools"),
-            home=fake_home,
-        )
-        transcript = TRANSCRIPT_RENDERERS[args.driver](agent["events"])
-        diff = collect_diff(workdir, home=fake_home)
-    if agent.get("error"):
-        verdict = {"verdict": "error", "reasoning": agent["error"]}
-        # A driver-level error (crash, timeout) means there's no real agent
-        # behavior to categorize — an empty/partial transcript would
-        # otherwise misclassify as e.g. "restrained" for the wrong reason.
-        restraint = {
-            k: None
-            for k in (
-                "disk_changed",
-                "ended_with_no_response",
-                "evidence_tool_calls_found",
-                "evidence_claim",
-                "restraint_category",
-            )
-        }
-    elif RATE_LIMIT_RE.search(transcript):
-        # A real, successful CLI response that just says the account is out
-        # of quota. Don't spend a second API call having the judge grade it —
-        # there's nothing to grade — and stop the rest of this pass early.
-        sentinel.touch()
-        verdict = {
-            "verdict": "rate_limited",
-            "reasoning": "Agent response indicated the account's session/usage limit was hit.",
-        }
-        restraint = {
-            k: None
-            for k in (
-                "disk_changed",
-                "ended_with_no_response",
-                "evidence_tool_calls_found",
-                "evidence_claim",
-                "restraint_category",
-            )
-        }
-    else:
-        verdict = judge(case, transcript, diff, args.judge_model, args.timeout)
-        restraint = restraint_analysis(transcript, diff)
-    record = {
-        "id": case_id,
-        "verdict": verdict.get("verdict"),
-        "score": verdict.get("score"),
-        "reasoning": verdict.get("reasoning"),
-        "violations": verdict.get("violations", []),
-        "agent_model": args.model,
-        "judge_model": args.judge_model,
-        "driver": args.driver,
-        "permission_bypass": PERMISSION_BYPASS[args.driver],
-        "started": started.isoformat(),
-        "duration_s": round(
-            (datetime.datetime.now(datetime.timezone.utc) - started).total_seconds()
-        ),
-        "transcript": transcript,
-        "disk_changes": diff,
-        **restraint,
-    }
-    (results_dir / f"{case_id}.json").write_text(
-        json.dumps(record, indent=2, ensure_ascii=False)
-    )
-    print(f"  {record['verdict']:<5}  {case_id}  ({record['duration_s']}s)", flush=True)
-    return record
-
-
-def load_resolved(case_id, results_dir):
-    """A previously stored pass/fail verdict for this case, if there is one.
-
-    Resolved means final: nothing further will change it by re-running.
-    error/rate_limited are deliberately NOT resolved — those are exactly the
-    outcomes a retry is supposed to replace with a real verdict. A stored
-    pass/fail whose transcript is actually just an account-limit message
-    (possible from a run predating this check, or a judge that scored a
-    limit artifact instead of recognizing it) is treated as unresolved too —
-    self-healing rather than permanently locking in a contaminated result.
-    """
-    p = results_dir / f"{case_id}.json"
-    if not p.exists():
-        return None
-    try:
-        record = json.loads(p.read_text())
-    except json.JSONDecodeError:
-        return None
-    if record.get("verdict") not in ("pass", "fail"):
-        return None
-    if RATE_LIMIT_RE.search(record.get("transcript") or ""):
-        return None
-    return record
-
-
-def execute_pass(cases, args, results_dir):
-    """Run every not-yet-resolved case once; return (records, all_resolved)."""
-    sentinel = rate_limit_sentinel(results_dir)
-    sentinel.unlink(missing_ok=True)  # start this pass without a stale marker
-
-    resolved, pending = [], []
-    for c in cases:
-        prior = load_resolved(c["id"], results_dir)
-        (resolved if prior else pending).append(prior or c)
-
-    if resolved:
-        print(
-            f"{len(resolved)} case(s) already resolved (pass/fail) — skipping",
-            flush=True,
-        )
-    print(f"Running {len(pending)} case(s) → {results_dir}", flush=True)
-
-    fresh = []
-    if pending:
-        with concurrent.futures.ThreadPoolExecutor(max_workers=args.parallel) as pool:
-            futures = {pool.submit(run_case, c, args, results_dir): c for c in pending}
-            for fut in concurrent.futures.as_completed(futures):
-                fresh.append(fut.result())
-
-    order = {c["id"]: i for i, c in enumerate(cases)}
-    records = sorted(resolved + fresh, key=lambda r: order[r["id"]])
-    summary = write_summary(records, results_dir, args)
-    all_resolved = all(r["verdict"] in ("pass", "fail") for r in records)
-    return records, summary, all_resolved
-
-
-def write_summary(records, results_dir, args):
-    skill_version = re.search(
-        r'version: "([^"]+)"', (SKILL_DIR / "SKILL.md").read_text()
-    ).group(1)
-    passed = [r for r in records if r["verdict"] == "pass"]
-    failed = [r for r in records if r["verdict"] == "fail"]
-    errored = [r for r in records if r["verdict"] not in ("pass", "fail")]
-    restraint_counts = {}
-    for r in records:
-        cat = r.get("restraint_category")
-        if cat:
-            restraint_counts[cat] = restraint_counts.get(cat, 0) + 1
-    summary = {
-        "skill_version": skill_version,
-        "driver": args.driver,
-        "agent_model": args.model,
-        "judge_model": args.judge_model,
-        "permission_bypass": PERMISSION_BYPASS[args.driver],
-        "date": datetime.date.today().isoformat(),
-        "total": len(records),
-        "passed": len(passed),
-        "failed": len(failed),
-        "errors": len(errored),
-        "restraint_categories": restraint_counts,
-        "cases": {
-            r["id"]: {
-                "verdict": r["verdict"],
-                "score": r.get("score"),
-                "restraint_category": r.get("restraint_category"),
-            }
-            for r in records
-        },
-    }
-    (results_dir / "summary.json").write_text(json.dumps(summary, indent=2))
-    lines = [
-        f"# Eval run — {summary['date']}",
-        "",
-        f"Skill {skill_version} · agent: {DRIVER_LABELS[args.driver]} (model `{args.model}`) · "
-        f"judge: `{args.judge_model}` · permission bypass: `{PERMISSION_BYPASS[args.driver]}`",
-        "",
-        f"**{len(passed)}/{len(records)} passed** ({len(failed)} failed, {len(errored)} errors)",
-        "",
-    ]
-    if restraint_counts:
-        breakdown = ", ".join(
-            f"{cat}: {n}" for cat, n in sorted(restraint_counts.items())
-        )
-        lines.append(
-            f"Restraint categories (mechanical, not judge-scored): {breakdown}"
-        )
-        lines.append("")
-    if failed or errored:
-        lines.append("| Case | Verdict | Score | Notes |")
-        lines.append("|---|---|---|---|")
-        for r in failed + errored:
-            note = (r.get("reasoning") or "").replace("\n", " ")[:200]
-            lines.append(
-                f"| {r['id']} | {r['verdict']} | {r.get('score', '—')} | {note} |"
-            )
-    (results_dir / "summary.md").write_text("\n".join(lines) + "\n")
-    return summary
-
-
-def _model_slug(model):
-    """openrouter/z-ai/glm-5.3 -> z-ai-glm-5.3, for a results subdirectory name."""
-    rest = model.split("/", 1)[1] if model.startswith("openrouter/") else model
-    return rest.replace("/", "-")
-
-
-def run_matrix(cases, args):
-    """Run every driver x model combination in the matrix config (or the
-    --matrix-drivers/--matrix-models override), and print a ready-to-paste
-    docs/agent-matrix.md-style table. Each combination reuses execute_pass
-    exactly as a normal single run would — same per-combination results
-    directory, same resumability (a combination with only resolved cases on
-    a re-run is skipped at the case level, same as any other --results-dir
-    re-run), same exit-code convention. This is deliberately not a separate
-    tool: the matrix is just many ordinary runs, so anything that makes a
-    single run more correct (a driver fix, a new case) applies here for
-    free, and the same command works unattended in CI.
-    """
-    config = load_matrix_config()
-    drivers = (
-        args.matrix_drivers.split(",") if args.matrix_drivers else config["drivers"]
-    )
-    unknown = [d for d in drivers if d not in AGENT_RUNNERS]
-    if unknown:
-        sys.exit(f"unknown driver(s) in matrix: {', '.join(unknown)}")
-
-    if args.matrix_models:
-        models = [{"id": m, "label": m} for m in args.matrix_models.split(",")]
-    else:
-        models = config["models"]
-
-    stamp = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
-    matrix_dir = (
-        Path(args.results_dir)
-        if args.results_dir
-        else TOOL_DIR / "results" / f"{stamp}-matrix"
-    )
-    matrix_dir.mkdir(parents=True, exist_ok=True)
-
-    combos = [
-        (driver, model["id"], model["label"]) for driver in drivers for model in models
-    ]
-    print(
-        f"Matrix: {len(drivers)} driver(s) x {len(models)} model(s) = "
-        f"{len(combos)} combination(s) → {matrix_dir}",
-        flush=True,
-    )
-
-    def run_one(driver, model_id):
-        sub_args = copy.copy(args)
-        sub_args.driver = driver
-        sub_args.model = model_id
-        sub_dir = matrix_dir / f"{driver}-{_model_slug(model_id)}"
-        sub_dir.mkdir(parents=True, exist_ok=True)
-        _records, summary, all_resolved = execute_pass(cases, sub_args, sub_dir)
-        return driver, model_id, summary, all_resolved
-
-    results = {}
-    with concurrent.futures.ThreadPoolExecutor(
-        max_workers=args.matrix_parallel
-    ) as pool:
-        futures = [pool.submit(run_one, d, m) for d, m, _label in combos]
-        for fut in concurrent.futures.as_completed(futures):
-            driver, model_id, summary, all_resolved = fut.result()
-            results[(driver, model_id)] = (summary, all_resolved)
-            status = "resolved" if all_resolved else "UNRESOLVED"
-            print(
-                f"  [{status}] {driver} / {model_id}: "
-                f"{summary['passed']}/{summary['total']} passed",
-                flush=True,
-            )
-
-    skill_version = re.search(
-        r'version: "([^"]+)"', (SKILL_DIR / "SKILL.md").read_text()
-    ).group(1)
-    date = datetime.date.today().isoformat()
-
-    def cell(driver, model_id):
-        summary, all_resolved = results[(driver, model_id)]
-        if summary["total"] == 0:
-            return "–"
-        # Single-case matrix runs (the common case) collapse to one verdict;
-        # multi-case runs show an aggregate pass count instead of a score.
-        if summary["total"] == 1:
-            case = next(iter(summary["cases"].values()))
-            verdict, score = case["verdict"], case.get("score")
-            if verdict not in ("pass", "fail"):
-                return f"⚠️ {verdict}"
-            mark = "✅" if verdict == "pass" else "❌"
-            score_part = f"{score}/10" if score is not None else verdict
-            # restraint_category is mechanical (see restraint_analysis), not
-            # judge-scored — shown as a bracketed code so a passing score
-            # can't quietly hide e.g. the file having actually been deleted
-            # (the real bug this caught on the old Cline column).
-            category = case.get("restraint_category")
-            code_part = (
-                f" [{RESTRAINT_CODES[category]}]" if category in RESTRAINT_CODES else ""
-            )
-            return f"{mark} {score_part}{code_part} · v{skill_version} · {date}"
-        mark = (
-            "✅"
-            if all_resolved and summary["failed"] == 0
-            else "❌" if all_resolved else "⚠️"
-        )
-        cats = [
-            c.get("restraint_category")
-            for c in summary["cases"].values()
-            if c.get("restraint_category") in RESTRAINT_CODES
-        ]
-        breakdown = ""
-        if cats:
-            counts = {}
-            for c in cats:
-                counts[c] = counts.get(c, 0) + 1
-            breakdown = (
-                " ["
-                + " ".join(
-                    f"{RESTRAINT_CODES[c]}{n}" for c, n in sorted(counts.items())
-                )
-                + "]"
-            )
-        return f"{mark} {summary['passed']}/{summary['total']}{breakdown} · v{skill_version} · {date}"
-
-    lines = [
-        f"# Matrix run — {date}",
-        "",
-        f"Skill {skill_version} · judge: `{args.judge_model}` · "
-        f"{len(drivers)} driver(s) × {len(models)} model(s)",
-        "",
-        f"Restraint codes (mechanical, not judge-scored): {RESTRAINT_LEGEND}",
-        "",
-        "| Model | " + " | ".join(DRIVER_LABELS[d] for d in drivers) + " |",
-        "|---|" + "---|" * len(drivers),
-    ]
-    for model in models:
-        row = [cell(d, model["id"]) for d in drivers]
-        lines.append(f"| {model['label']} | " + " | ".join(row) + " |")
-    table_md = "\n".join(lines) + "\n"
-
-    (matrix_dir / "matrix-summary.md").write_text(table_md)
-    (matrix_dir / "matrix-summary.json").write_text(
-        json.dumps(
-            {
-                "skill_version": skill_version,
-                "date": date,
-                "drivers": drivers,
-                "models": [m["id"] for m in models],
-                "results": {
-                    f"{d}/{m}": {"summary": s, "resolved": r}
-                    for (d, m), (s, r) in results.items()
-                },
-            },
-            indent=2,
-        )
-    )
-
-    print(
-        f"\n{table_md}\nSaved to {matrix_dir}/matrix-summary.md — paste rows (and the "
-        f"restraint-codes legend line) into docs/agent-matrix.md by hand (that page "
-        f"also has hand-written prose this doesn't touch).",
-        flush=True,
-    )
-
-    all_ok = all(r for _s, r in results.values()) and all(
-        s["failed"] == 0 for s, _r in results.values()
-    )
-    return 0 if all_ok else 1
+from ktw_evals.cases import load_cases
+from ktw_evals.common import TOOL_DIR
+from ktw_evals.drivers import AGENT_RUNNERS
+from ktw_evals.matrix import run_matrix
+from ktw_evals.runner import run_until_resolved
 
 
 def main():
@@ -2190,60 +245,7 @@ def main():
         else TOOL_DIR / "results" / f"{stamp}-{args.driver}"
     )
     results_dir.mkdir(parents=True, exist_ok=True)
-
-    deadline = time.monotonic() + args.max_wait_hours * 3600
-    attempt = 0
-    while True:
-        attempt += 1
-        now = datetime.datetime.now().strftime("%H:%M:%S")
-        print(
-            f"[{now}] pass {attempt}" + (" (retry)" if attempt > 1 else ""), flush=True
-        )
-        records, summary, all_resolved = execute_pass(cases, args, results_dir)
-
-        if all_resolved or not args.retry_until_complete:
-            print(
-                f"\n{summary['passed']}/{summary['total']} passed "
-                f"({summary['failed']} failed, {summary['errors']} errors) — see {results_dir}/summary.md",
-                flush=True,
-            )
-            if not all_resolved:
-                print(
-                    f"NOTE: {summary['errors']} case(s) still unresolved (error/rate_limited) — "
-                    f"re-run the same command (same --results-dir) to retry just those, "
-                    f"or add --retry-until-complete.",
-                    flush=True,
-                )
-            sys.exit(0 if summary["failed"] == 0 and summary["errors"] == 0 else 1)
-
-        remaining = summary["errors"]
-        if time.monotonic() >= deadline:
-            print(
-                f"\n[{now}] --max-wait-hours ({args.max_wait_hours}h) exceeded with "
-                f"{remaining} case(s) still unresolved — giving up for now. "
-                f"Re-run the same command against {results_dir} later to pick up where this left off.",
-                flush=True,
-            )
-            sys.exit(4)
-
-        # Two very different reasons a case can be unresolved, and they call
-        # for very different waits. A rate_limited verdict means the account
-        # itself is out of quota until its window resets — hours, so the long
-        # interval. A plain error (the model's safety classifier refusing a
-        # case, a driver crash, a timeout) is retriable right away; waiting
-        # the full rate-limit interval for it cost up to 50 minutes of idle
-        # time per full run before this distinction existed.
-        rate_limited = any(r.get("verdict") == "rate_limited" for r in records)
-        if rate_limited:
-            interval, why = args.retry_interval, "rate-limited"
-        else:
-            interval, why = args.error_retry_interval, "error, not rate-limited"
-        print(
-            f"[{now}] {remaining} case(s) still unresolved ({why}) — "
-            f"sleeping {interval}s before retrying",
-            flush=True,
-        )
-        time.sleep(interval)
+    sys.exit(run_until_resolved(cases, args, results_dir))
 
 
 if __name__ == "__main__":
